fix: consolidate ImageSelector and HoverArea
